PowerUps spawn. No interaction yet. Temp graphic
PowerUps will appear at regular intervals randomly above, below, or at the vertical center of the screen, and scroll from right to left, slower than the pipes do. Currently there is no interaction effect for the Bird when it touches a PowerUp, that will come soon. The PowerUps currently look like one of the frames of the bird, this will be replaced with a different graphic later.